# public/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PropertyDesk REST API",
    "version": "0.1.0",
    "description": "REST API PropertyDesk SaaS platforme za prodaju novogradnje.\n\nSvi endpointi vraćaju standardni envelope:\n```json\n{ \"data\": ..., \"meta\": { \"requestId\": \"...\", \"pagination?\": {...} } }\n```\n\nSve greške idu kroz:\n```json\n{ \"error\": { \"code\": \"...\", \"message\": \"...\", \"requestId\": \"...\", \"fieldErrors?\": { \"field\": [\"poruka\"] } } }\n```",
    "contact": {
      "name": "PropertyDesk Engineering",
      "url": "https://propertydesk.app",
      "email": "dev@propertydesk.app"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Trenutni host (same-origin — koristi za Try it out)"
    },
    {
      "url": "https://my.propertydesk.app",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-propertydesk.session_token",
        "description": "Better Auth session cookie (`cookiePrefix: propertydesk`).\nU produkciji: `__Secure-propertydesk.session_token` (HttpOnly, Secure, SameSite=Lax).\nU lokalnom dev-u (bez Secure): `propertydesk.session_token`.\n\nSwagger Authorize **ne može** da setuje HttpOnly cookie — za Try it out\nprvo se uloguj na `/sign-in` u istom browseru (isti origin), pa tek onda\npozovi zaštićene endpointe. Cookie se šalje automatski."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "NO_ACTIVE_ORGANIZATION",
                  "ORGANIZATION_ACCESS_DENIED",
                  "ORGANIZATION_SUSPENDED",
                  "ORGANIZATION_RESTRICTED",
                  "PLATFORM_ADMIN_REQUIRED",
                  "VALIDATION_ERROR",
                  "NOT_FOUND",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "METHOD_NOT_ALLOWED",
                  "INTERNAL_ERROR",
                  "BAD_REQUEST",
                  "NOT_IMPLEMENTED"
                ],
                "description": "Stabilan mašinski kod. Mobilni klijenti treba da ga koriste za logiku, ne `message`."
              },
              "message": {
                "type": "string",
                "description": "Ljudska poruka na srpskom, prikladna za prikaz krajnjem korisniku."
              },
              "requestId": {
                "type": "string",
                "format": "uuid",
                "description": "Korelacioni ID za logove (istovetan `x-request-id` header-u)."
              },
              "fieldErrors": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "description": "Po polju — mapirane Zod greške. Samo za 422 VALIDATION_ERROR."
              }
            }
          }
        }
      },
...
    },
...
    "responses": {
      "Unauthenticated": {
        "description": "401 — sesija ne postoji ili je istekla.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "UNAUTHENTICATED",
                "message": "Morate biti prijavljeni.",
                "requestId": "f1d2a5f2-0e4a-4b13-8a51-96ecffec5d51"
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "403 — nemaš dozvolu za ovu akciju.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "FORBIDDEN",
                "message": "Nemate dozvolu za ovu akciju.",
                "requestId": "f1d2a5f2-0e4a-4b13-8a51-96ecffec5d51"
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/reports/export": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "List / read reports",
        "description": "**Auth:** `requirePermission(\"report.read\")`\n",
        "responses": {
          "200": {
            "description": "OK\n"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  }
}